requirements.txt
paho-mqtt==1.6.1
folium==0.16.0
python-dotenv==1.0.1
branca==0.7.2
requests==2.31.0
pynmea2==1.3.1
pyubx2==1.2.50
Pillow==10.2.0
selenium==4.16.0
webdriver-manager==4.0.1
pyserial==3.5
future==1.0.0